# assets/openapi.yaml
openapi: 3.1.0
info:
    title: Phishing API
    summary: Una base de datos con miles enlaces maliciosos.
    description: Una base de datos colaborativa con enlaces maliciosos centrada pero no dedicada en la moderación de Discord.
    termsOfService: 'https://discusss.gitbook.io/phishing/informacion'
    contact:
      name: 'Discusss'
      url: 'https://discord.gg/SJqrGP7'
      email: 'admin@lacabra.app'
    license:
      name: GPL-3.0
      url: 'https://www.gnu.org/licenses/gpl-3.0.html'
    version: 0.0.1-dev
...
servers:
  - url: 'https://phishing.lacabra.app/api'
    description: 'Servidor principal'
paths:
...
  /keys:
    get:
      summary: 'Obtener información sobre una clave de acceso'
      description: 'Obtener información sobre una clave de acceso dada su clave'
      operationId: 'getKey'
      tags:
        - 'Key'
      parameters:
        - name: 'key'
          in: 'query'
          description: 'Clave de acceso a obtener'
          required: true
          allowEmptyValue: false
          schema:
            type:
              - string
      security:
        - masterAuth: []
      responses:
        '200':
          description: 'Clave obtenida'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 'Estado de la petición (200)'
                  data:
                    type: object
                    description: 'Información de la clave'
                    properties:
                      key:
                        type: string
                        description: 'Clave de acceso'
                      created_at:
                        type: string
                        description: 'Fecha de creación de la clave, en formato RFC 3339'
                      expires_at:
                        type: integer
                        description: 'Fecha de expiración de la clave, en formato RFC 3339'
                      last_used_at:
                        type: string
                        description: 'Fecha de la última vez que se ha usado la clave, en formato RFC 3339'
                      owner:
                        type: string
                        description: 'Nombre del usuario dueño de la clave'
                      uses:
                        type: integer
                        description: 'Veces que se ha usado la clave'
                      ips:
                        type: array
                        description: 'IPs desde las que se ha usado la clave'
                        items:
                          type: string
                          description: 'IP'
                      user_agent:
                        type: string
                        description: 'User-Agent desde el que se ha usado la clave por última vez'
                      created_by:
                        type: string
                        description: 'Nombre del usuario que ha creado la clave'
                      notes:
                        type: string
                        description: 'Notas sobre la clave'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: 'Clave no encontrada'
          content:
            application/json:
              schema:
                type: object
                properties:
                status:
                  type: string
                  description: 'Estado de la petición (404)'
                message:
                  type: string
                  description: 'Mensaje de error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: 'Añadir una clave de acceso'
      description: 'Añadir una clave de acceso'
      operationId: 'addKey'
      tags:
        - 'Key'
      security:
        - masterAuth: [ ]
      requestBody:
        description: 'Datos de la clave a añadir'
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expires_at
                - owner
                - created_by
              properties:
                expires_at:
                  type: string
                  description: 'Fecha de expiración de la clave, en formato RFC 3339. Máximo 5 año'
                owner:
                  type: string
                  description: 'Nombre del usuario dueño de la clave'
                created_by:
                  type: integer
                  description: 'Nombre del usuario que ha creado la clave'
                notes:
                  type: string
                  description: 'Notas sobre la clave'
      responses:
        '200':
          description: 'Clave añadida'
          content:
            application/json:
              schema:
                type: object
                properties:
                status:
                  type: string
                  description: 'Estado de la petición (200)'
                data:
                  type: object
                  description: 'Clave añadida'
                  properties:
                    key:
                      type: string
                      description: 'Clave de acceso'
                    created_at:
                      type: string
                      description: 'Fecha de creación de la clave, en formato RFC 3339'
                    expires_at:
                      type: integer
                      description: 'Fecha de expiración de la clave, en formato RFC 3339'
                    owner:
                      type: string
                      description: 'Nombre del usuario dueño de la clave'
                    created_by:
                      type: string
                      description: 'Nombre del usuario que ha creado la clave'
                    notes:
                      type: string
                      description: 'Notas sobre la clave'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: 'Modificar una clave de acceso'
      description: 'Modificar una clave de acceso con su clave'
      operationId: 'patchKey'
      tags:
        - 'Key'
      security:
        - masterAuth: [ ]
      parameters:
        - name: 'key'
          in: 'path'
          description: 'Clave de acceso a modificar'
          required: true
          allowEmptyValue: false
          schema:
            type:
              - string
      requestBody:
        description: 'Datos de la clave a modificar'
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expires_at:
                  type: string
                  description: 'Fecha de expiración de la clave, en formato RFC 3339. Máximo 5 año'
                last_used_at:
                  type: integer
                  description: 'Fecha de la última vez que se ha usado la clave, en formato RFC 3339'
                owner:
                  type: string
                  description: 'Nombre del usuario dueño de la clave'
                uses:
                  type: integer
                  description: 'Veces que se ha usado la clave'
                ips:
                  type: array
                  description: 'IPs desde las que se ha usado la clave'
                  items:
                    type: string
                    description: 'IP'
                user_agent:
                  type: string
                  description: 'User-Agent desde el que se ha usado la clave por última vez'
                created_by:
                  type: string
                  description: 'Nombre del usuario que ha creado la clave'
                notes:
                  type: integer
                  description: 'Notas sobre la clave'
      responses:
        '200':
          description: 'Clave modificada'
          content:
            application/json:
              schema:
                type: object
                properties:
                status:
                  type: string
                  description: 'Estado de la petición (200)'
                data:
                  type: object
                  description: 'Dominio modificado'
                  properties:
                    expires_at:
                      type: string
                      description: 'Fecha de expiración de la clave, en formato RFC 3339. Máximo 5 año, solo si se ha modificado'
                    last_used_at:
                      type: integer
                      description: 'Fecha de la última vez que se ha usado la clave, en formato RFC 3339, solo si se ha modificado'
                    owner:
                      type: string
                      description: 'Nombre del usuario dueño de la clave, solo si se ha modificado'
                    uses:
                      type: integer
                      description: 'Veces que se ha usado la clave, solo si se ha modificado'
                    ips:
                      type: array
                      description: 'IPs desde las que se ha usado la clave, solo si se ha modificado'
                      items:
                        type: string
                        description: 'IP'
                    user_agent:
                      type: string
                      description: 'User-Agent desde el que se ha usado la clave por última vez, solo si se ha modificado'
                    created_by:
                      type: string
                      description: 'Nombre del usuario que ha creado la clave, solo si se ha modificado'
                    notes:
                      type: integer
                      description: 'Notas sobre la clave, solo si se ha modificado'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: 'Clave no encontrada'
          content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      description: 'Estado de la petición (404)'
                    message:
                      type: string
                      description: 'Mensaje de error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: 'Eliminar una clave de acceso'
      description: 'Eliminar una clave de acceso con su clave'
      operationId: 'deleteKey'
      tags:
        - 'Key'
      security:
        - masterAuth: [ ]
      parameters:
        - name: 'key'
          in: 'path'
          description: 'Clave de acceso a eliminar'
          required: true
          allowEmptyValue: false
          schema:
            type:
              - string
      responses:
        '200':
          description: 'Clave eliminada'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 'Estado de la petición (200)'
                  data:
                    type: object
                    description: 'Clave eliminada'
                    properties:
                      key:
                        type: string
                        description: 'Clave de acceso'
                      created_at:
                        type: string
                        description: 'Fecha de creación de la clave, en formato RFC 3339'
                      expires_at:
                        type: integer
                        description: 'Fecha de expiración de la clave, en formato RFC 3339'
                      last_used_at:
                        type: string
                        description: 'Fecha de la última vez que se ha usado la clave, en formato RFC 3339'
                      owner:
                        type: string
                        description: 'Nombre del usuario dueño de la clave'
                      uses:
                        type: integer
                        description: 'Veces que se ha usado la clave'
                      ips:
                        type: array
                        description: 'IPs desde las que se ha usado la clave'
                        items:
                          type: string
                          description: 'IP'
                      user_agent:
                        type: string
                        description: 'User-Agent desde el que se ha usado la clave por última vez'
                      created_by:
                        type: string
                        description: 'Nombre del usuario que ha creado la clave'
                      notes:
                        type: string
                        description: 'Notas sobre la clave'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: 'Clave no encontrada'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 'Estado de la petición (404)'
                  message:
                    type: string
                    description: 'Mensaje de error'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: 'Token de autenticación de usuario'
    masterAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: 'Token de autenticación de administrador, solo para uso interno'
  responses:
    UnauthorizedError:
      description: 'No autorizado'
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: 'Estado de la petición (401)'
              message:
                type: string
                description: 'Mensaje de error'
    InternalServerError:
      description: 'Error interno del servidor'
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: 'Estado de la petición (500)'
              message:
                type: string
                description: 'Mensaje de error'
    BadRequestError:
      description: 'Petición incorrecta o vacía'
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: 'Estado de la petición (400)'
              message:
                type: string
                description: 'Mensaje de error'
...